+++ src/app/(auth)/(reg)/register/page.tsx
@@ -3,20 +3,37 @@
 import { useState } from 'react'
 import { useRouter } from 'next/navigation'
 import { useRegFormContext } from '@/contexts/RegFormContext'
+import { initialRegFormData } from '@/constants/regFormData'
+import { RegFormData } from '@/types/regFormData'
 
 export default function RegisterPage() {
-  const [email, setEmail] = useState('')
-  const [password, setPassword] = useState('')
+  const [registerForm, setRegisterForm] =
+    useState<RegFormData>(initialRegFormData)
+  const [invalidFormMessage, setInvalidFormMessage] = useState('')
+
   const router = useRouter()
   const { setRegFormData } = useRegFormContext()
 
   const handleSubmit = (e: React.FormEvent) => {
     e.preventDefault()
-    console.log(email, password)
-    if (email && password) {
-      setRegFormData({ email, password })
-      router.replace('/verify/verify-email')
+
+    if (registerForm.phoneNumber && registerForm.password) {
+      setRegFormData(registerForm)
+      router.replace('/verify/verify-phone')
+      // router.replace('/verify/verify-email')
     }
+  }
+
+  const handleChange = (
+    e: React.ChangeEvent<HTMLInputElement | HTMLSelectElement>
+  ) => {
+    const { id, type } = e.target
+    // const value = type === "checkbox" ? e.target.checked : e.target.value;
+    if (invalidFormMessage) {
+      setInvalidFormMessage('')
+    }
+
+    setRegisterForm((prev) => ({ ...prev, [id]: e.target.value }))
   }
 
   return (
@@ -32,8 +49,20 @@ export default function RegisterPage() {
           <input
             type="email"
             id="email"
-            value={email}
-            onChange={(e) => setEmail(e.target.value)}
+            value={registerForm.email}
+            onChange={handleChange}
+            className="w-full px-3 py-2 border rounded"
+          />
+        </div>
+        <div className="mb-4">
+          <label className="block text-gray-700 mb-2" htmlFor="email">
+            Телефон
+          </label>
+          <input
+            type="phone"
+            id="phoneNumber"
+            value={registerForm.phoneNumber}
+            onChange={handleChange}
             className="w-full px-3 py-2 border rounded"
             required
           />
@@ -45,8 +74,8 @@ export default function RegisterPage() {
           <input
             type="password"
             id="password"
-            value={password}
-            onChange={(e) => setPassword(e.target.value)}
+            value={registerForm.password}
+            onChange={handleChange}
             className="w-full px-3 py-2 border rounded"
             required
           />