+++ src/lib/auth-client.ts
@@ -1,5 +1,12 @@
-import { createAuthClient } from "better-auth/react";
+import { phoneNumberClient } from 'better-auth/client/plugins'
+import { createAuthClient } from 'better-auth/react'
 
 export const authClient = createAuthClient({
-  baseURL: "http://localhost:3000",
-});
+  plugins: [phoneNumberClient()]
+})
+
+// import { createAuthClient } from "better-auth/react";
+
+// export const authClient = createAuthClient({
+//   baseURL: "http://localhost:3000",
+// });